Emulators API

The Emulators submodule defines types and methods for emulating expensive SimulatorForwardProblems.

SimulationBasedInference.Emulators.Decorrelated — Type

Decorrelated{bijType} <: DataTransform

Decorrelation transform applied to a multivariate output (target) space.

SimulationBasedInference.Emulators.EmulatedObservables — Type

EmulatedObservables{names} <: SciMLBase.AbstractSciMLAlgorithm

Represents an emulated forward solver for observables names in a given forward problem. The dimensionality of the output spaces of the emulators are assumed to match those of the observables exactly.

SimulationBasedInference.Emulators.Emulator — Type

Emulator{TM,TT} <: Emulator

Data structure consisting of some training data for a model emulator, appropriate transforms, and a tuple of univariate regressors which are applied to the transformed data.

SimulationBasedInference.Emulators.EmulatorData — Type

EmulatorData

Generic container for emulator training data matrices X and Y. X should have shape m x N where N is the number of samples and m is the number of covariates. Y should have dimensions N x d where d is the number of ouptut covariates.+

SimulationBasedInference.Emulators.GPRegressor — Type

GPRegressor

Generic implementation of a Gaussian Process regressor for a univariate outputs. This implementation is adapted from MLJGaussianProcesses to be usable without MLJ.

SimulationBasedInference.Emulators.StackedRegressors — Type

StackedRegressors{TM}

Simple representation of a set of stacked, univaraite regressors for multi-target regression problems.